requirements.txt
absl-py==0.12.0
aiohttp==3.7.4.post0
alabaster==0.7.12
altgraph==0.17.2
appdirs==1.4.4
APScheduler==3.6.3
asgiref==3.5.0
astunparse==1.6.3
async-timeout==3.0.1
asyncio==3.4.3
attrs==21.2.0
auto-py-to-exe==2.18.0
Automat==20.2.0
Babel==2.9.1
beautifulsoup4==4.9.3
bleach==4.1.0
boto3==1.18.36
botocore==1.21.36
bottle==0.12.19
bottle-websocket==0.2.9
breadability==0.1.20
bs4==0.0.1
cachetools==4.2.2
certifi==2021.10.8
cffi==1.14.6
characteristic==14.3.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.1
cloudpickle==1.6.0
colorama==0.4.4
coloredlogs==15.0.1
comtypes==1.1.10
ConfigArgParse==1.5.2
constantly==15.1.0
cssselect==1.1.0
cx-Freeze==6.10
cx-Logging==3.0
cycler==0.10.0
defusedxml==0.7.1
discord==1.0.1
discord-py-slash-command==3.0.1
discord.py==1.7.3
dislash.py==1.4.9
Django==4.0.2
docopt==0.6.2
docutils==0.17.1
easyocr==1.4.1
Eel==0.12.4
elote==0.1.0
entrypoints==0.3
fake-useragent==0.1.11
fakeredis==1.6.0
fbmessenger==4.3.1
feedparser==6.0.8
filelock==3.0.12
Flask==2.0.1
flatbuffers==1.12
future==0.18.2
gast==0.4.0
get-all-tickers==1.8
gevent==21.8.0
gevent-websocket==0.10.1
google==3.0.0
google-auth==1.30.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
graphviz==0.17
greenlet==1.1.1
grpcio==1.34.1
h11==0.12.0
h5py==3.1.0
httpcore==0.12.3
httpx==0.16.1
huggingface-hub==0.0.8
humanfriendly==9.2
hyperlink==21.0.0
idna==3.3
imageio==2.9.0
imagesize==1.2.0
importlib-metadata==4.11.3
incremental==21.3.0
instant-api-client==0.1.2
ipython-genutils==0.2.0
itsdangerous==2.0.1
jaraco.context==4.0.0
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.0.1
jsonpickle==2.0.0
jsonschema==3.2.0
jupyter-client==7.0.2
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
keras==2.6.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
klein==21.8.0
lief==0.11.5
lxml==4.8.0
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
mistune==0.8.4
more-itertools==8.7.0
MouseInfo==0.1.3
mpl-finance==0.10.1
mplfinance==0.12.7a17
multidict==5.1.0
multitasking==0.0.9
mysql-connector-python==8.0.31
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nbsphinx==0.8.7
nest-asyncio==1.5.1
networkx==2.6.2
newsapi-python==0.2.6
nltk==3.6.2
nose==1.3.7
numerize==0.12
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.4.60
opencv-python-headless==4.5.1.48
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.4
pandas-datareader==0.9.0
pandoc==1.1.0
pandocfilters==1.4.3
parse==1.19.0
pathlib==1.0.1
pefile==2021.9.3
Pillow==9.0.1
pipreqs==0.4.11
plotly==5.0.0
plumbum==1.7.0
ply==3.11
prettytable==3.5.0
protobuf==3.17.2
prsaw==0.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycountry==20.7.3
pycparser==2.20
pyee==8.1.0
pygame==2.0.1
PyGetWindow==0.0.9
Pygments==2.10.0
pyinstaller==4.10
pyinstaller-hooks-contrib==2022.2
pykwalify==1.6.0
pymongo==3.11.4
PyMsgBox==1.0.9
pyodbc==4.0.31
pyparsing==2.4.7
pyperclip==1.8.2
pypiwin32==223
pyppeteer==0.2.5
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyquery==1.4.3
pyreadline==2.1
PyRect==0.2.0
pyrsistent==0.18.0
PyScreeze==0.1.28
PySide2==5.15.2
python-bidi==0.4.2
python-dateutil==2.8.1
python-docx==0.8.11
python-dotenv==0.19.0
python-telegram-bot==13.7
pytweening==1.0.4
pytz==2021.1
PyWavelets==1.3.0
pywin32==301
pywin32-ctypes==0.2.0
pywinauto==0.5.1
PyYAML==5.4.1
pyzmq==22.2.1
rasa-core==0.8.6
rasa-nlu==0.11.5
redis==3.5.3
regex==2021.4.4
requests==2.27.1
requests-html==0.10.0
requests-oauthlib==1.3.0
rfc3986==1.5.0
rsa==4.7.2
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
s3transfer==0.5.0
sacremoses==0.0.45
schedule==1.1.0
scikit-image==0.19.2
scikit-learn==0.24.2
scipy==1.6.3
selenium==3.141.0
sgmllib3k==1.0.0
shiboken2==5.15.2
simplejson==3.17.2
six==1.15.0
sklearn==0.0
slackclient==2.9.3
sniffio==1.2.0
snowballstemmer==2.1.0
sortedcontainers==2.4.0
soupsieve==2.2.1
Sphinx==4.1.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.2
sumy==0.8.1
tdqm==0.0.1
tenacity==7.0.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
testpath==0.5.0
threadpoolctl==2.1.0
tifffile==2022.2.9
tkcap==0.0.2
tokenizers==0.10.3
torch==1.11.0
torchvision==0.12.0
tornado==6.1
tqdm==4.60.0
traitlets==5.1.0
transformers==4.6.1
ttkwidgets==0.12.1
Tubes==0.2.0
Twisted==21.7.0
twisted-iocpsupport==1.0.2
typing-extensions==3.7.4.3
tzdata==2021.5
tzlocal==2.1
urllib3==1.26.9
w3lib==1.22.0
wcwidth==0.2.5
webencodings==0.5.1
websockets==8.1
Werkzeug==2.0.1
whichcraft==0.6.1
wikipedia==1.4.0
wolframalpha==5.0.0
wrapt==1.12.1
xlrd==2.0.1
xmltodict==0.12.0
yahoo-fin==0.8.8
yahoo-finance==1.4.0
yarg==0.1.9
yarl==1.6.3
yfinance==0.1.63
zipp==3.7.0
zope.event==4.5.0
zope.interface==5.4.0
